Server Custom Exceptions

+++ java/src/main/java/com/techelevator/service/RestMusicService.java
@@ -1,10 +1,13 @@
 package com.techelevator.service;
 
 import com.techelevator.dao.MusicDao;
+import com.techelevator.exceptions.MusicAlreadyExistsException;
+import com.techelevator.exceptions.MusicNotFoundException;
 import com.techelevator.model.Music;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class RestMusicService implements MusicService {
@@ -19,24 +22,49 @@ public class RestMusicService implements MusicService {
         return musicDao.listAllMusicSuggestions();
     }
 
-    public Music musicSuggestionByMusicId(Integer musicId) { return musicDao.getMusicSuggestionByMusicId(musicId); }
+    public Music musicSuggestionByMusicId(Integer musicId) {
+        try {
+            Optional<Music> musicData = Optional.ofNullable(musicDao.getMusicSuggestionByMusicId(musicId));
+            if (musicData.isEmpty()) {
+                throw new MusicNotFoundException("Music Suggestion Not Found!");
+            }
+        }
+        catch (MusicNotFoundException mnfe){
+            System.out.println(mnfe.getMessage());
+        }
 
-    public Music musicSuggestionByArtistAndSong(String artist, String song) { return musicDao.getMusicSuggestionByArtistAndSong(artist, song); }
+        return musicDao.getMusicSuggestionByMusicId(musicId);
+    }
+
+    public Music musicSuggestionByArtistAndSong(String artist, String song) {
+        try {
+            Optional<Music> musicData = Optional.ofNullable(musicDao.getMusicSuggestionByArtistAndSong(artist, song));
+            if (musicData.isEmpty()) {
+                throw new MusicNotFoundException("Music Suggestion Not Found! Incorrect Artist and/or Song");
+            }
+        }
+        catch (MusicNotFoundException mnfe){
+            System.out.println(mnfe.getMessage());
+        }
+
+        return musicDao.getMusicSuggestionByArtistAndSong(artist, song);
+    }
+
     public boolean create(Music musicSuggestion) {
         boolean isCreated = false;
 
         try {
             if (musicDao.getMusicSuggestionByArtistAndSong(musicSuggestion.getArtist(), musicSuggestion.getSong()).getArtist().equals(musicSuggestion.getArtist()) && musicDao.getMusicSuggestionByArtistAndSong(musicSuggestion.getArtist(), musicSuggestion.getSong()).getSong().equals(musicSuggestion.getSong())) {
-                System.out.println("Already Exists");
+                throw new MusicAlreadyExistsException("Music Suggestion Already Exists");
             } else {
                 musicDao.createMusicSuggestion(musicSuggestion);
                 isCreated = true;
             }
         }
-        catch (Exception e){
-            System.out.println(e.getMessage());
+        catch (MusicAlreadyExistsException maee){
+            System.out.println(maee.getMessage());
         }
-        
+
         return isCreated;
     }
 
@@ -44,11 +72,16 @@ public class RestMusicService implements MusicService {
         boolean isUpdated = false;
 
         try {
-            musicDao.updateMusicSuggestion(musicSuggestion, musicId);
-            isUpdated = true;
+            Optional<Music> musicData = Optional.ofNullable(musicDao.getMusicSuggestionByMusicId(musicId));
+            if (musicData.isEmpty()) {
+                throw new MusicNotFoundException("Music Suggestion Not Found! Unable to update.");
+            } else {
+                musicDao.updateMusicSuggestion(musicSuggestion, musicId);
+                isUpdated = true;
+            }
         }
-        catch (Exception e){
-            System.out.println(e.getMessage());
+        catch (MusicNotFoundException mnfe){
+            System.out.println(mnfe.getMessage());
         }
 
         return isUpdated;
@@ -58,11 +91,16 @@ public class RestMusicService implements MusicService {
         boolean isDeleted = false;
 
         try {
-            musicDao.deleteMusicSuggestion(musicId);
-            isDeleted = true;
+            Optional<Music> musicData = Optional.ofNullable(musicDao.getMusicSuggestionByMusicId(musicId));
+            if (musicData.isEmpty()) {
+                throw new MusicNotFoundException("Music Suggestion Not Found! Unable to delete.");
+            } else {
+                musicDao.deleteMusicSuggestion(musicId);
+                isDeleted = true;
+            }
         }
-        catch (Exception e){
-            System.out.println(e.getMessage());
+        catch (MusicNotFoundException mnfe){
+            System.out.println(mnfe.getMessage());
         }
 
         return isDeleted;
